refactor: try-with-resources 적용
