+++ app/src/main/java/com/techcourse/support/context/ContextLoaderListener.java
@@ -1,7 +1,7 @@
 package com.techcourse.support.context;
 
 import com.techcourse.config.DataSourceConfig;
-import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
+import com.techcourse.support.jdbc.init.DatabasePopulatingUtils;
 import jakarta.servlet.ServletContextEvent;
 import jakarta.servlet.ServletContextListener;
 import jakarta.servlet.annotation.WebListener;
@@ -11,6 +11,6 @@ public class ContextLoaderListener implements ServletContextListener {
 
     @Override
     public void contextInitialized(ServletContextEvent sce) {
-        DatabasePopulatorUtils.execute(DataSourceConfig.getInstance());
+        DatabasePopulatingUtils.execute(DataSourceConfig.getInstance());
     }
 }
