--- app/src/main/java/com/techcourse/support/jdbc/init/DatabasePopulatorUtils.java
@@ -1,49 +1,0 @@
-package com.techcourse.support.jdbc.init;
-
-import java.io.File;
-import java.net.URL;
-import java.nio.file.Files;
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.sql.Statement;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class DatabasePopulatorUtils {
-
-    private static final Logger LOG = LoggerFactory.getLogger(DatabasePopulatorUtils.class);
-
-    private DatabasePopulatorUtils() {
-    }
-
-    public static void execute(DataSource dataSource) {
-        Connection connection = null;
-        Statement statement = null;
-        try {
-            final URL url = DatabasePopulatorUtils.class.getClassLoader().getResource("schema.sql");
-            final File file = new File(url.getFile());
-            final String sql = Files.readString(file.toPath());
-
-            connection = dataSource.getConnection();
-            statement = connection.createStatement();
-            statement.execute(sql);
-        } catch (Exception e) {
-            LOG.error(e.getMessage(), e);
-        } finally {
-            try {
-                if (statement != null) {
-                    statement.close();
-                }
-            } catch (SQLException ignored) {
-            }
-
-            try {
-                if (connection != null) {
-                    connection.close();
-                }
-            } catch (SQLException ignored) {
-            }
-        }
-    }
-}
