+++ app/src/test/java/com/techcourse/dao/UserDaoTest.java
@@ -4,9 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.techcourse.config.DataSourceConfig;
 import com.techcourse.domain.User;
-import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
+import com.techcourse.support.jdbc.init.DatabasePopulatingUtils;
 import java.util.List;
-import nextstep.jdbc.JdbcTemplate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -16,7 +15,7 @@ class UserDaoTest {
 
     @BeforeEach
     void setup() {
-        DatabasePopulatorUtils.execute(DataSourceConfig.getInstance());
+        DatabasePopulatingUtils.execute(DataSourceConfig.getInstance());
         userDao = new UserDao(DataSourceConfig.getInstance());
 
         final User user = new User("gugu", "password", "hkkang@woowahan.com");
